refactor(canvas): #373 Phase 4-3 CanvasLayout.tsx を 3 hook/utility に分割
Issue #373 / Phase 4-3: `layouts/CanvasLayout.tsx` (896 行) から pure helper /
チーム復元 / auto-save の 3 責務を別 module / hook に切り出し。
move only / 機能変更なし / 不変式 #5 (Canvas 常時マウント) 維持。

# 構成

```
layouts/CanvasLayout.tsx                       896 → ~665 行 (UI orchestration + JSX)
lib/canvas-layout-helpers.ts             新規   70 行  localeOf / formatCardCount /
                                                      formatAgentCount / mergeCanvasMembers /
                                                      serializeAutoSavePayload (pure)
lib/hooks/use-canvas-team-restore.ts     新規   82 行  Issue #159: setupTeamMcp 復元
                                                      (in_flight / failed / done state machine)
lib/hooks/use-canvas-auto-save.ts        新規  146 行  Issue #167 / #132 / #124: debounce
                                                      auto-save (saveBatch + drag-bailout)
```

# 不変式の維持

- **Canvas 常時マウント (#5)**: `style={isCanvasActive ? undefined : { display: 'none' }}` の
  display:none 切替は touch せず。display:none → unmount に変えると PTY が kill される
- 機能変更なし: チーム復元の 30 秒 backoff / auto-save 1500ms debounce / mergeCanvasMembers の
  sessionId キュー追跡など全てのロジックを bit identical で hook 化
- Issue #167 (recent ref 経由), Issue #132 (saveBatch), Issue #124 (drag bailout は caller 側) の
  経緯コメントも逐字保持

# 検証結果

- `npm run typecheck`: ✅ green

# Agent Team で並列分担

3 hook/utility の draft を 3 人の teammate (canvas-helpers / canvas-team-restore
/ canvas-auto-save extractor) に並列で書かせ、team-lead 側で CanvasLayout.tsx
の不要 import 削除 + hook 呼び出しに置換 (62 + 105 行を 8 行に圧縮) + typecheck
通しを実施。

Refs #373
